edgeadmin/requirements.txt
grpcio==1.35.0
protobuf==3.14.0
grpcio-tools==1.35.0
opencv-python==4.7.0.68
requests==2.28.2